moved all components into folders
