+++ client/components/Cart/Checkout.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import {checkout} from '../store/cart'
-import Checkouts from '../../react-express-stripe/frontend/Checkout'
+import {checkout} from '../../store/cart'
+import Checkouts from '../../../react-express-stripe/frontend/Checkout'
 
 function Checkout(props) {
   return (
