+++ client/components/index.js
@@ -4,6 +4,6 @@
  * having to figure out which file they belong to!
  */
 export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
+export {default as UserHome} from './User/UserHome'
 export {Login, Signup} from './auth-form'
-export {AllNoodles} from './AllNoodles'
+export {AllNoodles} from './Noodle/AllNoodles'
