+++ client/store/index.js
@@ -15,9 +15,11 @@ const reducer = combineReducers({
   noodles: noodlesReducer,
   noodle: noodleReducer
 })
+
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
+
 const store = createStore(reducer, middleware)
 
 export default store
